+++ src/config/redux/constants/SLICE_NAMES.ts
@@ -1,4 +1,5 @@
 export default {
+  auth: 'auth',
   error: 'error',
   offers: 'offers',
   cities: 'cities',
