+++ src/config/redux/slice/auth/selector.ts
@@ -1,0 +1,5 @@
+import { State } from '../..';
+
+export const selectAuthState = (s: State) => s.auth;
+
+export const selectAuthStatus = (s: State) => selectAuthState(s).status;
