+++ src/config/redux/slice/auth/state.ts
@@ -1,0 +1,9 @@
+export interface AuthSliceState {
+  status: boolean;
+}
+
+export function getEmptyState(): AuthSliceState {
+  return {
+    status: false,
+  };
+}
