+++ src/domain/auth/hooks/useAuth.ts
@@ -1,3 +1,6 @@
-export function useAuth() {
-  return true;
+import { useSelector } from 'react-redux';
+import { selectAuthStatus } from '../../../config/redux/slice/auth/selector';
+
+export function useAuthStatus() {
+  return useSelector(selectAuthStatus);
 }
