+++ src/domain/router/components/PrivateRoute.tsx
@@ -1,8 +1,8 @@
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../../auth';
+import { useAuthStatus } from '../../auth';
 import routes from '../constants/ROUTES';
 
 export function PrivateRoute({ children }: { children: JSX.Element }) {
-  const auth = useAuth();
+  const auth = useAuthStatus();
   return auth ? children : <Navigate to={routes.login} />;
 }
