+++ src/domain/ui/components/Header.tsx
@@ -1,0 +1,65 @@
+import { Link } from 'react-router-dom';
+import routes from '../../router/constants/ROUTES';
+import { useAuthStatus } from '../../auth/hooks/useAuth';
+
+export function Header() {
+  const isAuth = useAuthStatus();
+
+  return (
+    <header className="header">
+      <div className="container">
+        <div className="header__wrapper">
+          <div className="header__left">
+            <Link
+              className="header__logo-link header__logo-link--active"
+              to={routes.cities}
+            >
+              <img
+                className="header__logo"
+                src="img/logo.svg"
+                alt="6 cities logo"
+                width="81"
+                height="41"
+              />
+            </Link>
+          </div>
+          <nav className="header__nav">
+            {isAuth ? (
+              <ul className="header__nav-list">
+                <li className="header__nav-item user">
+                  <Link
+                    className="header__nav-link header__nav-link--profile"
+                    to={routes.favorites}
+                  >
+                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
+                    <span className="header__user-name user__name">
+                      Oliver.conner@gmail.com
+                    </span>
+                    <span className="header__favorite-count">3</span>
+                  </Link>
+                </li>
+                <li className="header__nav-item">
+                  <a className="header__nav-link" href="#">
+                    <span className="header__signout">Sign out</span>
+                  </a>
+                </li>
+              </ul>
+            ) : (
+              <ul className="header__nav-list">
+                <li className="header__nav-item user">
+                  <a
+                    className="header__nav-link header__nav-link--profile"
+                    href="#"
+                  >
+                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
+                    <span className="header__login">Sign in</span>
+                  </a>
+                </li>
+              </ul>
+            )}
+          </nav>
+        </div>
+      </div>
+    </header>
+  );
+}
