+++ src/pages/Main.tsx
@@ -1,5 +1,5 @@
 import { Offer } from '../domain/offer/types';
-import { Header } from '../domain/ui/layout/Header';
+import { Header } from '../domain/ui/components/Header';
 import { Map } from '../domain/map';
 import { useMemo, useState } from 'react';
 import { Navbar } from '../domain/city/components/Navbar';
